requirements.txt
paramiko>=3.0.0
PyYAML>=6.0
pexpect>=4.8.0
